finished abstracted validated input out of guesser
